refactor: Move google login btn folder
